split demo components
